Removed keybind service as it was cleaner to merge it with the keyboard service
