--- src/Sidekick.Common.Platform/IKeybindProvider.cs
@@ -1,16 +1,0 @@
-using Sidekick.Common.Initialization;
-using Sidekick.Common.Keybinds;
-
-namespace Sidekick.Common.Platform
-{
-    /// <summary>
-    /// Service providing keybind functions
-    /// </summary>
-    public interface IKeybindProvider : IInitializableService
-    {
-        /// <summary>
-        /// A list of keybind handlers currently registered to the application
-        /// </summary>
-        Dictionary<string, IKeybindHandler> KeybindHandlers { get; }
-    }
-}
