+++ src/Sidekick.Common.Platform/Keyboards/KeyboardProvider.cs
@@ -1,10 +1,13 @@
 using System.Diagnostics;
 using System.Text;
 using System.Text.RegularExpressions;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using SharpHook;
 using SharpHook.Native;
 using Sidekick.Common.Initialization;
+using Sidekick.Common.Keybinds;
 
 namespace Sidekick.Common.Platform.Keyboards
 {
@@ -130,7 +133,8 @@ namespace Sidekick.Common.Platform.Keyboards
         private static readonly Regex ModifierKeys = new("Ctrl|Shift|Alt");
 
         private readonly ILogger<KeyboardProvider> logger;
-        private readonly IKeybindProvider keybindProvider;
+        private readonly IOptions<SidekickConfiguration> configuration;
+        private readonly IServiceProvider serviceProvider;
         private readonly IProcessProvider processProvider;
 
         private bool HasInitialized { get; set; } = false;
@@ -140,15 +144,19 @@ namespace Sidekick.Common.Platform.Keyboards
 
         public KeyboardProvider(
             ILogger<KeyboardProvider> logger,
-            IKeybindProvider keybindProvider,
+            IOptions<SidekickConfiguration> configuration,
+            IServiceProvider serviceProvider,
             IProcessProvider processProvider)
         {
             this.logger = logger;
-            this.keybindProvider = keybindProvider;
+            this.configuration = configuration;
+            this.serviceProvider = serviceProvider;
             this.processProvider = processProvider;
         }
 
         public event Action<string>? OnKeyDown;
+
+        private Dictionary<string, IKeybindHandler> KeybindHandlers { get; init; } = new();
 
         /// <inheritdoc/>
         public InitializationPriority Priority => InitializationPriority.Low;
@@ -162,13 +170,24 @@ namespace Sidekick.Common.Platform.Keyboards
                 return Task.CompletedTask;
             }
 
+            // Initialize keybindings
+            KeybindHandlers.Clear();
+            foreach (var keybindType in configuration.Value.Keybinds)
+            {
+                var keybindHandler = (IKeybindHandler)serviceProvider.GetRequiredService(keybindType);
+                foreach (var keybind in keybindHandler.GetKeybinds())
+                {
+                    KeybindHandlers.Add(keybind, keybindHandler);
+                }
+            }
+
+            // Initialize keyboard hook
             Hook = new();
-
             Hook.KeyPressed += OnKeyPressed;
-
             HookTask = Hook.RunAsync();
             Simulator = new EventSimulator();
 
+            // Make sure we don't run this multiple times
             HasInitialized = true;
 
             return Task.CompletedTask;
@@ -203,7 +222,7 @@ namespace Sidekick.Common.Platform.Keyboards
             var keybind = str.ToString();
             OnKeyDown?.Invoke(keybind);
 
-            if (processProvider.IsPathOfExileInFocus && keybindProvider.KeybindHandlers.TryGetValue(keybind, out var keybindHandler))
+            if (processProvider.IsPathOfExileInFocus && KeybindHandlers.TryGetValue(keybind, out var keybindHandler))
             {
                 if (keybindHandler.IsValid())
                 {
