+++ src/Sidekick.Common.Platform/StartupExtensions.cs
@@ -2,7 +2,6 @@ using System.Runtime.InteropServices;
 using Microsoft.Extensions.DependencyInjection;
 using Sidekick.Common.Platform.Clipboard;
 using Sidekick.Common.Platform.Keyboards;
-using Sidekick.Common.Platform.Options;
 using Sidekick.Common.Platform.Windows.Processes;
 using Sidekick.Common.Platforms.Localization;
 
@@ -30,7 +29,6 @@ namespace Sidekick.Common.Platform
                 services.AddSidekickInitializableService<IProcessProvider, ProcessProvider>();
             }
 
-            services.AddSidekickInitializableService<IKeybindProvider, KeybindProvider>();
             services.AddSidekickInitializableService<IKeyboardProvider, KeyboardProvider>();
 
             return services;
